+++ web/stores/flows.ts
@@ -170,8 +170,7 @@ export const useFlowsStore = defineStore('flowsStore', {
 						task_id: task_id,
 						flow_name: task.name,
 						progress: task.progress,
-						input_prompt: <string>task.input_params?.prompt || '',
-						seed: <string>task.input_params?.seed || '',
+						input_files: task.input_files || [],
 						input_params_mapped: task.input_params || null,
 						error: task?.error || null,
 						outputs: task.outputs,
@@ -517,9 +516,22 @@ export const useFlowsStore = defineStore('flowsStore', {
 							return
 						}
 						runningFlow.progress = <number>progress[task_id].progress
+						// progress[task_id].error = 'Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... Test error message... '
 						if (progress[task_id].error && progress[task_id].error !== '') {
 							runningFlow.error = progress[task_id].error
 							return
+						}
+						if (progress[task_id].execution_time) {
+							runningFlow.execution_time = progress[task_id].execution_time
+						}
+						// update input_params_mapped with new values
+						runningFlow.input_params_mapped = {
+							...runningFlow.input_params_mapped,
+							...progress[task_id].input_params,
+						}
+						// update input_files
+						if (progress[task_id].input_files) {
+							runningFlow.input_files = progress[task_id].input_files
 						}
 						if (progress[task_id].progress === 100) {
 							// Remove finished flow from running list
@@ -691,9 +703,11 @@ export interface FlowRunning {
 	task_id: string
 	flow_name: string
 	progress: number
+	input_files?: TaskInputFile
 	input_params_mapped: TaskHistoryInputParam
 	outputs: FlowOutputParam[]
 	error?: string
+	execution_time?: number
 }
 
 export interface FlowProgress {
@@ -719,6 +733,15 @@ export interface TasksHistory {
 	[task_id: string]: TaskHistoryItem
 }
 
+export interface TaskInputFile {
+	[index: number]: TaskInputFileData
+}
+
+export interface TaskInputFileData {
+	file_name: string
+	file_size: string
+}
+
 export interface TaskHistoryInputParam {
 	[name: string]: any
 }
@@ -726,6 +749,7 @@ export interface TaskHistoryInputParam {
 export interface TaskHistoryItem {
 	name: string
 	input_params: TaskHistoryInputParam
+	input_files: TaskInputFile
 	progress: number
 	error?: string
 	outputs: FlowOutputParam[]